Expose Block/Grid Dim to specialize IR

+++ lib/JitEngineDeviceCUDA.cpp
@@ -82,7 +82,7 @@ std::unique_ptr<MemoryBuffer> JitEngineDeviceCUDA::extractDeviceBitcode(
 }
 
 void JitEngineDeviceCUDA::setLaunchBoundsForKernel(Module &M, Function &F,
-                                                   int GridSize,
+                                                   size_t GridSize,
                                                    int BlockSize) {
   NamedMDNode *NvvmAnnotations = M.getNamedMetadata("nvvm.annotations");
   assert(NvvmAnnotations && "Expected non-null nvvm.annotations metadata");
